backend/requirements.txt
openai-whisper
google-generativeai
requests
sounddevice
soundfile
schedule
numpy
setuptools
python-dotenv